+++ devicemanagement/tests/TestDevMgmtApiSyncCmds.h
@@ -239,8 +239,10 @@ protected:
   void testStateInspectionReadCSR(uint64_t shireID, uint64_t threadMask, uint64_t hartID, uint64_t csrName);
   void testStateInspectionWriteCSR(uint64_t shireID, uint64_t threadMask, uint64_t hartID, uint64_t csrName,
                                    uint64_t csrData);
-  void readMem(uint64_t readAddr);
-  void writeMem(uint64_t testInputData, uint64_t writeAddr);
+  void readMem_unprivileged(uint64_t readAddr);
+  void readMem_privileged(uint64_t readAddr);
+  void writeMem_unprivileged(uint64_t testInputData, uint64_t writeAddr);
+  void writeMem_privileged(uint64_t testInputData, uint64_t writeAddr);
 
   inline Target getTestTarget(void) const {
     auto envTarget = getenv("TARGET");