+++ devicemanagement/tests/mdi_debug/StateInspectionApiTestcmds.cpp
@@ -33,18 +33,36 @@ class StateInspectionApiTestcmds : public TestDevMgmtApiSyncCmds {
   }
 };
 
-TEST_F(StateInspectionApiTestcmds, readMem) {
+TEST_F(StateInspectionApiTestcmds, readMem_unprivileged) {
   if (targetInList({Target::Silicon, Target::SysEMU})) {
-    readMem(COMPUTE_KERNEL_DEVICE_ADDRESS);
+    readMem_unprivileged(COMPUTE_KERNEL_DEVICE_ADDRESS);
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;
   }
 }
 
-TEST_F(StateInspectionApiTestcmds, writeMem) {
+TEST_F(StateInspectionApiTestcmds, readMem_privileged) {
   if (targetInList({Target::Silicon, Target::SysEMU})) {
-    writeMem(MDI_TEST_WRITE_MEM_TEST_DATA, COMPUTE_KERNEL_DEVICE_ADDRESS);
+    readMem_privileged(MM_FW_MASTER_SDATA_BASE);
+  } else {
+    DV_LOG(INFO) << "Skipping the test since its not supported on current target";
+    FLAGS_enable_trace_dump = false;
+  }
+}
+
+TEST_F(StateInspectionApiTestcmds, writeMem_unprivileged) {
+  if (targetInList({Target::Silicon, Target::SysEMU})) {
+    writeMem_unprivileged(MDI_TEST_WRITE_MEM_TEST_DATA, COMPUTE_KERNEL_DEVICE_ADDRESS);
+  } else {
+    DV_LOG(INFO) << "Skipping the test since its not supported on current target";
+    FLAGS_enable_trace_dump = false;
+  }
+}
+
+TEST_F(StateInspectionApiTestcmds, writeMem_privileged) {
+  if (targetInList({Target::Silicon, Target::SysEMU})) {
+    writeMem_privileged(MDI_TEST_WRITE_MEM_TEST_DATA, MM_FW_MASTER_SDATA_BASE);
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;
@@ -53,8 +71,7 @@ TEST_F(StateInspectionApiTestcmds, writeMem) {
 
 TEST_F(StateInspectionApiTestcmds, testStateInspectionReadGPR) {
   if (targetInList({Target::Silicon, Target::SysEMU})) {
-    testStateInspectionReadGPR(MDI_TEST_DEFAULT_SHIRE_ID, MDI_TEST_DEFAULT_THREAD_MASK,
-                               MDI_TEST_DEFAULT_HARTID);
+    testStateInspectionReadGPR(MDI_TEST_DEFAULT_SHIRE_ID, MDI_TEST_DEFAULT_THREAD_MASK, MDI_TEST_DEFAULT_HARTID);
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;
@@ -63,8 +80,8 @@ TEST_F(StateInspectionApiTestcmds, testStateInspectionReadGPR) {
 
 TEST_F(StateInspectionApiTestcmds, testStateInspectionWriteGPR) {
   if (targetInList({Target::Silicon, Target::SysEMU})) {
-    testStateInspectionWriteGPR(MDI_TEST_DEFAULT_SHIRE_ID, MDI_TEST_DEFAULT_THREAD_MASK,
-                                MDI_TEST_DEFAULT_HARTID, MDI_TEST_GPR_WRITE_TEST_DATA);
+    testStateInspectionWriteGPR(MDI_TEST_DEFAULT_SHIRE_ID, MDI_TEST_DEFAULT_THREAD_MASK, MDI_TEST_DEFAULT_HARTID,
+                                MDI_TEST_GPR_WRITE_TEST_DATA);
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;
@@ -73,8 +90,8 @@ TEST_F(StateInspectionApiTestcmds, testStateInspectionWriteGPR) {
 
 TEST_F(StateInspectionApiTestcmds, testStateInspectionReadCSR) {
   if (targetInList({Target::Silicon, Target::SysEMU})) {
-    testStateInspectionReadCSR(MDI_TEST_DEFAULT_SHIRE_ID, MDI_TEST_DEFAULT_THREAD_MASK,
-                               MDI_TEST_DEFAULT_HARTID, MDI_TEST_CSR_PC_REG);
+    testStateInspectionReadCSR(MDI_TEST_DEFAULT_SHIRE_ID, MDI_TEST_DEFAULT_THREAD_MASK, MDI_TEST_DEFAULT_HARTID,
+                               MDI_TEST_CSR_PC_REG);
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;
@@ -83,8 +100,8 @@ TEST_F(StateInspectionApiTestcmds, testStateInspectionReadCSR) {
 
 TEST_F(StateInspectionApiTestcmds, testStateInspectionWriteCSR) {
   if (targetInList({Target::Silicon, Target::SysEMU})) {
-    testStateInspectionWriteCSR(MDI_TEST_DEFAULT_SHIRE_ID, MDI_TEST_DEFAULT_THREAD_MASK,
-                                MDI_TEST_DEFAULT_HARTID, MDI_TEST_CSR_PC_REG, MDI_TEST_CSR_WRITE_PC_TEST_ADDRESS);
+    testStateInspectionWriteCSR(MDI_TEST_DEFAULT_SHIRE_ID, MDI_TEST_DEFAULT_THREAD_MASK, MDI_TEST_DEFAULT_HARTID,
+                                MDI_TEST_CSR_PC_REG, MDI_TEST_CSR_WRITE_PC_TEST_ADDRESS);
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;